feat(fee-calculation): implemented logic

+++ tezos-rpc/src/models/limits.rs
@@ -2,8 +2,8 @@ use num_bigint::BigUint;
 
 pub const GAS_SAFETY_MARGIN: u64 = 100;
 pub const STORAGE_SAFETY_MARGIN: u64 = 100;
-pub const FEE_SAFTY_MARGIN: u64 = 100;
-pub const BASE_FEE: u64 = 100;
+pub const FEE_SAFTY_MARGIN: u8 = 100;
+pub const BASE_FEE: u8 = 100;
 pub const FEE_PER_GAS_UNIT: u64 = 100; // nanoTez
 pub const FEE_PER_STORAGE_BYTE: u64 = 1000; // nanoTez
 pub const FEE_PER_PAID_STORAGE_SIZE_DIFF_BYTE: u64 = 250; // mutez
